Plot visibility, B/W option, minor bug fixes
Changed color scheme for plots away from jet to something that hopefully
shows up better. Symbols for presence locations are now different. B/W
hopefully easier to see. Boundaries and large text added to suitability
plot. Feedback and an input prompt for color or B/W added to the
plotting procedure. Fixed an error in the readme description, and an
error by which caught IOError messages were not printed.